requirements.txt
streamlit
langchain-anthropic
langchain-tavily
langchain-community
python-dotenv
faiss-cpu
langchain-huggingface
sentence-transformers
fpdf